# Cache derivative computation in direct solver