--- a/appveyor.yml
+++ b/appveyor.yml
@@ -43,7 +43,9 @@ install:
   - rmdir C:\\cygwin /s /q
 
   # Update previous packages and install the build and runtime dependencies of the project.
-  - conda update --all --yes
+  # XXX: setuptools>23 is currently broken on Win+py3 with numpy
+  # (https://github.com/pypa/setuptools/issues/728)
+  - conda update --all --yes setuptools=23
   - conda install --quiet --yes numpy scipy cython nose scikit-learn wheel"
   - conda install --quiet --yes -c conda-forge sklearn-contrib-lightning
   # there seems to be a problem with conda-build 1.21.0 on python27 and win64, avoid this