@vene 2016-08-13
The Appveyor Python 3.5 errors seem to be caused by a bug in the new setuptools release (25.1.6). I experienced the same build error on my Windows box.