@vene 2016-08-14
Merged by rebase. Fixed #4 in the process.